QE-12617 Make has_text matcher always have higher priority (#352)
In one of the tests, we noticed that when looking for an element named
`Version`, fuzzy find put an element named `Version 1` first. This is
caused by the fact that fuzzy find has two parts and each part does
`has_text` matcher first and `contains` matcher second. Henceforth, the
`contains` matcher of the first part has higher priority of the
`has_text` matcher of the second part.

This PR puts the two parts together such that it does `has_text` matcher
first and then `contains` matcher.