QE-13274 Enable color output in logs when running in parallel (#364)
Currently when running in parallel, the cucu logs html files don't have
color. With this PR, they will.

The main change is in `src/cucu/cli/run.py` and
`src/cucu/formatter/cucu.py`. The rest of the change is to use
`--no-color-output` in tests so that the `STDOUT` won't have color marks