QE-13684 Fix injecting cucu fuzzy find library (#372)
In cases where a webpage redirects to another webpage, the context of
the browser is cleared and the loaded javascript is lost. This causes
`jQuery` or `jqCucu` not defined error.

We don't need to load the needed javascript code at the beginning of the
new browser since it's loaded at each `fuzzy.find` call.

In this PR, the injection of the all the needed JS code is combined into
one browser execution. This reduces the chance of half loaded code