QE-13818 Helper template for 2 interacting elements, added draggables (#381)
### Helper function

I was looking to implement steps in cucu for draggable elements and I
looked a little deeper at the helper code and found that there is a
potential enhancement that cucu could have, templatized steps for 2
interacting elements.

I started by implementing draggables as a proof of concept as a custom
step on e2e-test branch. I then transferred that code over to cucu
framework and had it working without any helper function implementation.

That same code is now embedded in a helper function called
`define_interaction_on_thing_with_name_steps` (final name pending)
basically this is what it does... It intertwines 3 functions.

1. find function for finding the element you want to make the action
with
2. find function for the finding the element you want the above element
to interact with
3. Action function that defines what happens between the two found
elements, in my case I found a draggable element, I found another
generic element and I performed a drag action between them.

It will be easiest to start on the `src/cucu/steps/draggable_steps.py`
and then see how the helper function from here `src/cucu/helpers.py` is
being used to allow for modularity across different applications.

### Draggable function

After going through some trials on different browsers I noticed that the
implementations for ActionChains was not consistent enough to be used in
cucu. I then pivoted and created a JS_DRAG_DROP global variable that is
then referenced in the `drag_element_to_element` function.

This took a bit of refining to get right. Waits needs to be put in place
to prevent the browser from executing too fast for the Webdriver to
realize anything has happened. And with that we need to stay synchronous
with python code that is being executed in tandem. That is where the
Webdriver wait comes in.

I setup a Webdriver wait to keep python executions and javascript
executions in sync.

`WebDriverWait(driver, 10).until(lambda driver:
driver.execute_script("return window.dragAndDropCompleted;"))`

This allows the drag and drop to be very consistent across all browser
types.