QE-12356 Add border to elements we expected to see (#373)
This is a proof of concept for highlighting elements that were the
target of an "expect to see" step.

If deemed that this is a good idea, then we can proceed to highlight
other things, as well. I'm not sure what the next target would be,
though; things we actually interact with tend to produce changes in the
page that either already indicate which element the test tried to
interact with, or that rerender the element and make it stale.