QE-14257 Fix jQuery loading in regex steps (#408)
regex steps load jQuery again and if it follows a fuzzy find step and
then is followed by another fuzzy find step, the second fuzzy find step
won't use the correct jQuery. This PR makes sure that if jQuery is
already loaded, the regex steps won't reload it.

A functional test is created to reproduce the issue and verify the fix.