QE-14478 Only obfuscate certain places in json (#414)
We have a case where a secret is a number and cucu obfuscates the
timestamp value that causes failure to generate cucu report.

The places where secrets may be leaked are in step variables, step
outputs and error messages. Therefore, instead of search and replace the
whole json stream, cucu should only look for these three places.