QE-14935 Improve clear_input (#434)
`input_.clear()` in `clear_input()` sometimes causes an element not
interactable. Since the function has another method of clearing the
input, `input_.clear()` is not needed.

Surprisingly, removing the call uncovers another bug in cucu, which is
that sending a key to the input accidentally clears the existing
content. This is not desired since the most use case for that step is
after writing to the input, we then send an enter key to trigger an
action. This PR fixes the behavior.

The change is also tested in E2E tests