QE-16383 Add report path in JUnit when available (#513)
## PR Type
What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->
- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting)
- [ ] Refactoring (no functional changes)
- [ ] CI related changes
- [ ] Other... Please describe:

## What is the current behavior?
When html report is generated, JUnit file doesn't have the path and it's
up to `domino-testrail` reporter to construct the path. Therefore,
whenever we change the path, we need to update `domino-testrail`

Issue Number: https://dominodatalab.atlassian.net/browse/QE-16383

## What is the new behavior?
When html report is generated, the path is added to the JUnit file and
`domino-testrail` can get the info directly.

## Does this PR introduce a breaking change?
- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information